added floor file

+++ 2-floor.py
@@ -1,8 +1,10 @@
 #!/usr/bin/env python3
 """ floor function """
 
+import math
 
-def floor(n: float) -> float:
+
+def floor(n: float) -> int:
     """
     write a annotated function
     """
